lab1_2/backend/requirements.txt
djangorestframework
django-filter
django-cors-headers
faker
markdown
psycopg[binary,pool]
gunicorn